Update project to adhere to linting rules
