--- app/src/main/java/org/squidfish/tuple_n_back/MainActivity.kt
@@ -1,21 +1,0 @@
-package org.squidfish.tuple_n_back
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import org.squidfish.tuple_n_back.ui.theme.TupleNbackTheme
-import org.squidfish.tuple_n_back.views.TupleNBackApp
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-
-        setContent {
-            TupleNbackTheme {
-                TupleNBackApp()
-            }
-        }
-    }
-}
